Wheelchair - fixed wireless timeout regression bug, changed sensitivity numbers, improved dead zone algorithms

+++ wheelchair/src/menu.h
@@ -19,7 +19,7 @@ float menuGetTurnThrow(void);
 uint8_t menuGetTopFwdSpeed(void);
 uint8_t menuGetTopRevSpeed(void);
 uint8_t menuGetTopTurnSpeed(void);
-float menuGetSensitivity(void);
+double menuGetSensitivity(void);
 uint8_t menuGetAcceleration(void);
 uint8_t menuGetDeceleration(void);
 uint8_t menuGetOuterDeadBand(void);
